requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
opencv-python==4.8.1.78
numpy<2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=4.0.0
httpx==0.25.2
tqdm==4.66.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
passlib>=1.7.4
python-jose==3.3.0
bcrypt==3.2.2
Pillow==10.1.0
pyjwt==2.10.1
stripe>=8.0.0
slowapi
python-multipart
email-validator
alembic
psutil==5.9.6
imagehash==4.3.1
paramiko==3.4.0

# Redis caching
redis==4.6.0
fastapi-cache2[redis]==0.2.1

# PDF generation
reportlab==4.0.7
PyPDF2==3.0.1

# AI Tagging
openai>=1.50.0  # Optional - only needed for AI-powered tagging
pytesseract>=0.3.10  # OCR for free text extraction from images

# Email Messaging
sendgrid>=6.11.0  # SendGrid API for transactional and marketing emails
schedule>=1.2.0  # Background scheduler for scheduled email campaigns